fixup! Refactor Vagrantfile generation via string.Template
