+++ kiwi/storage/subformat/vagrant_base.py
@@ -18,108 +18,16 @@
 import json
 import os
 import random
-from textwrap import dedent
+
 from tempfile import mkdtemp
-from string import Template
 
 from kiwi.storage.subformat.base import DiskFormatBase
+from kiwi.storage.subformat.template.vagrant_config import VagrantConfigTemplate
 from kiwi.command import Command
 
 from kiwi.exceptions import (
     KiwiFormatSetupError
 )
-
-
-class VagrantConfigTemplate(object):
-    """
-    **Generate a Vagrantfile configuration template**
-
-    This class creates a simple template for the Vagrantfile that is included
-    inside a vagrant box.
-
-    The included Vagrantfile carries additional information for vagrant: by
-    default that is just the boxes' MAC address, but depending on the provider
-    additional information need to be present. These can be passed via the
-    parameter ``custom_settings`` to the method :meth:`get_template`.
-
-    Example usage:
-
-    The default without any additional settings will result in this
-    Vagrantfile:
-
-    .. code:: python
-
-        >>> vagrant_config = VagrantConfigTemplate()
-        >>> print(
-        ...     vagrant_config.get_template()
-        ...     .substitute({'mac_address': 'deadbeef'})
-        ... )
-        Vagrant.configure("2") do |config|
-          config.vm.base_mac = "deadbeef"
-        end
-
-    If your provider/box requires additional settings, provide them as follows:
-
-    .. code:: python
-
-        >>> extra_settings = dedent('''
-        ... config.vm.hostname = "no-dead-beef"
-        ... config.vm.provider :special do |special|
-        ...   special.secret_settings = "please_work"
-        ... end
-        ... ''').strip()
-        >>> print(
-        ...     vagrant_config.get_template(extra_settings)
-        ...     .substitute({'mac_address': 'DEADBEEF'})
-        ... )
-        Vagrant.configure("2") do |config|
-          config.vm.base_mac = "DEADBEEF"
-          config.vm.hostname = "no-dead-beef"
-          config.vm.provider :special do |special|
-            special.secret_settings = "please_work"
-          end
-        end
-    """
-
-    _INDENT = '  '
-
-    _HEADER = dedent('''
-        Vagrant.configure("2") do |config|
-    ''').strip() + os.linesep
-
-    _MAC = _INDENT + dedent('''
-        config.vm.base_mac = "${mac_address}"
-    ''').strip() + os.linesep
-
-    _END = dedent('''
-        end
-    ''').strip()
-
-    @classmethod
-    def get_template(cls, custom_settings=None):
-        """
-        Return a new template with ``custom_settings`` included and
-        indented appropriately.
-
-        :param str custom_settings: String of additional settings that get pasted
-            into the Vagrantfile template. The string is put at the correct
-            indentation level for you, but the internal indentation has to be
-            provided by the caller.
-        :return: A template with ``custom_settings`` inserted at the
-            appropriate position. The template has one the variable
-            ``mac_address`` that must be substituted.
-        :rtype: string.Template
-        """
-        template_data = cls._HEADER
-        template_data += cls._MAC
-        if custom_settings:
-            template_data += cls._INDENT
-            template_data += cls._INDENT.join(
-                custom_settings.splitlines(True)
-            )
-            template_data += os.linesep
-        template_data += cls._END
-        return Template(template_data)
 
 
 class DiskFormatVagrantBase(DiskFormatBase):
